resources/aws_route53_record: Fix set_identifier with underscore
As mentioned in detail in https://github.com/terraform-providers/terraform-provider-aws/issues/11677
& https://github.com/terraform-providers/terraform-provider-aws/issues/6298
there was a bug if we supplied a `set_identifier` which should contain an
underscore in.
This fix changes the logic by searching the first occurrences of underscores
to calculate the correct values for each field and as a result for the
`set_identifier` as well.
The tests included proves that the change doesn't break something and also
fixes the aforementioned bug.